+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 ###### [ 1.0.18 ] - 2023/11/15
 
+  *  Added VS test class `testVS_SoundTouchDiscovery.py` to test Zeroconf discovery scenarios.
+  *  Updated VS test class `testVS_SoundTouchClient` with missing testing scenarios.
+
+###### [ 1.0.18 ] - 2023/11/15
+
   *  Added method to `SoundTouchClient`: AddMusicServiceSources - Adds any servers in the `MediaServerList` to the sources list if they do not exist in the sources list as a "STORED_MUSIC" source.
   *  Added method to `SoundTouchClient`: RemoveMusicServiceAccount - Removes an existing music service account from the sources list.
   *  Added method to `SoundTouchClient`: SetMusicServiceAccount - Adds a music service account to the sources list.